Fix for ceylon/ceylon-compiler#706: Nested class not found at runtime
We now remember which classes do NOT satisfy basic or identifiable, rather than
look up classes by name, which does not work, or rely on the metamodel, which is slow.
Since the classes that do not satisfy basic/identifiable are in the overwhelming minority,
this seems like the best choice.