Add test for #550: parseFloat("-0.5")
Moves the parseFloat() tests to a separate function for better
organisation. Also adds a helper function to extract the common
'(parseFloat("...") else 0.0)==...' pattern (because parseFloat() has an
optional return type, which cannot be used with '==' directly), and
changes the "fallback" value from 0.0 to NaN (which will always make the
comparison fail).