module lookup fallback to search by ClassLoader
It's possible for module lookup by package name to fail in at least two cases:

1. Classes that are created at runtime, such as jdk dynamic proxies have the
   package name com.sun.proxy

2. Classes that were available at compile time, but not associated with a
   module at runtime due to a less controlled environment, such as a servlet
   container

Of course, this patch will not help in case #2 for classes in loaders that do
not have associated modules.

Perhaps a better (but more involved) fix would be to dynamically create a new
package and classloader specific module when no suitable module exists.