Fix String.firstIndexWhere() for last index
The method tested the index into the underlying string against the
length right after incrementing the index, so the last character was
always skipped. A simpler fix for this would have been to move the check
(either to the beginning of the loop or the end), but instead I changed
the loop condition and put the `return null;` at the end, where it
logically belongs. I think it's nicer this way.